+++ src/cloze.rs
@@ -189,6 +189,7 @@ pub fn expand(
             image_back: None,
             frontend: None,
             max_stage: None,
+            at: None,
         });
     }